[BE] add: Debate 완료

+++ BackEnd/pjt1/src/main/java/moonrise/pjt1/debate/repository/DebateRepository.java
@@ -3,5 +3,5 @@ package moonrise.pjt1.debate.repository;
 import moonrise.pjt1.debate.entity.DebateEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DebateRepository extends JpaRepository<DebateEntity, Integer> {
+public interface DebateRepository extends JpaRepository<DebateEntity, Long> {
 }
